画像ロードを高速化 (存在しないファイルに対してloadBitmapを呼び出さないように改善)
git-svn-id: http://www.mz3.jp/svn/mz3/trunk@1219 875e8a4c-fb2d-0410-984e-92a9dd373959

+++ src/MZ3/MZ3BackgroundImage.h
@@ -6,6 +6,8 @@
  */
 #pragma once
 
+#include "util_base.h"
+
 /**
  * ”wŒi—p‰æ‘œŠÇ—ƒNƒ‰ƒX
  */
@@ -52,6 +54,9 @@ public:
 
 	bool load( LPCTSTR szFilepath )
 	{
+		if (!util::ExistFile(szFilepath)) {
+			return false;
+		}
 		if (!m_hBitmap) {
 			deleteBitmap();
 		}
